+++ src/main/java/com/vybz/busker_info_service/busker_sns/application/BuskerSnsService.java
@@ -1,8 +1,7 @@
 package com.vybz.busker_info_service.busker_sns.application;
 
-import com.vybz.busker_info_service.busker_sns.dto.request.RequestAddBuskerSnsDto;
+import com.vybz.busker_info_service.busker_sns.dto.request.RequestUpsertBuskerSnsDto;
 import com.vybz.busker_info_service.busker_sns.dto.request.RequestDeleteBuskerSnsDto;
-import com.vybz.busker_info_service.busker_sns.dto.request.RequestUpdateBuskerSnsDto;
 import com.vybz.busker_info_service.busker_sns.dto.response.ResponseBuskerSnsDto;
 
 import java.util.List;
@@ -10,22 +9,16 @@ import java.util.List;
 public interface BuskerSnsService {
 
     /**
-     * 버스커 sns 생성
-     * @param requestAddBuskerSnsDto
+     * 버스커 sns 생성/수정
+     * @param requestUpsertBuskerSnsDto
      */
-    void createBuskerSns(RequestAddBuskerSnsDto requestAddBuskerSnsDto);
+    void upsertBuskerSns(RequestUpsertBuskerSnsDto requestUpsertBuskerSnsDto);
 
     /**
      * 유저 uuid로 버스커 sns 조회
      * @param buskerUuid
      */
     List<ResponseBuskerSnsDto> findBuskerSnsByBuskerUuid(String buskerUuid);
-
-    /**
-     * 버스커 sns 업데이트
-     * @param requestUpdateBuskerSnsDto
-     */
-    void updateBuskerSns(RequestUpdateBuskerSnsDto requestUpdateBuskerSnsDto);
 
     /**
      * 버스커 sns 삭제
